Created connect() to handle Context.Consumer
Abstracted the store connection logic away to the connect function, so a user can simply run connect to link a Component with the Redux Store.

Connect returns a function that you immediately invoke by passing the Component Name as an argument.

Connect takes a function as an argument. It should return an object matching the state the Component will use.